nydusd: move chunk v1 related tests into mod v1
Move chunk v1 related tests into mod v1.
